Unify `FilteredQuerySource` and `SelectStatement`
`FilteredQuerySource` only existed to make sure you couldn't pass a
query with a select clause or any other clause besides `WHERE` as a
target for `update` or `delete`. Instead of introducing a parallel
query type, we can instead just add the notion of a "default" select
clause.

I intentionally wanted this to be a voldemort type, so I needed to
rearrange the parameters on `SelectStatement` which caused a bit more
churn than I would have liked.

Close #760.